refactor(framework): Move k3d harness under dev k8s
