+++ framework/dev/k8s/assets/probe_app/launch_probe/server_app.py
@@ -1,0 +1,22 @@
+# Copyright 2026 Flower Labs GmbH. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+import flwr as fl
+
+app = fl.serverapp.ServerApp()
+
+
+@app.main()
+def main(grid, context):
+    print("K8s launch probe ServerApp ran")
