+++ framework/dev/k8s/assets/probe_app/pyproject.toml
@@ -1,0 +1,19 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "launch-probe"
+version = "0.1.0"
+
+[tool.hatch.build.targets.wheel]
+packages = ["launch_probe"]
+
+[tool.flwr.app]
+publisher = "flwrlabs"
+
+[tool.flwr.app.components]
+serverapp = "launch_probe.server_app:app"
+clientapp = "launch_probe.client_app:app"
+
+[tool.flwr.app.config]
